Адаптация под новый клиент базы
В новой версии OrientDB.NET изменились способы доступа к базе
Также исправлена работа с новым API Вконтакте. Система возвращена к
работоспособному состоянию

+++ VKAggregator/Statistic/Statistic.cs
@@ -85,6 +85,11 @@ namespace VKAggregator.Statistic
 
         public static void LogEnd()
         {
+            if (!Directory.Exists("log"))
+            {
+                DirectoryInfo di = Directory.CreateDirectory("log");
+                return;
+            }
             endTime = DateTime.Now;
             File.Create("log/log_" + DateTime.Now.Day + "_" + DateTime.Now.Month + ".log");
             using (StreamWriter file =
